Added debug prints (and finally found a potential bug with bullets, hurray!), also tested project on VS and fixed stuff to make it work there

+++ Emitter.cpp
@@ -13,5 +13,5 @@ std::array<short, 2> Emitter::getEmitterAbsolutePos()
 {
     std::array<short, 2> parentPos = parentObj->getPos();
 
-    return {parentPos[0] + rel_X, parentPos[1] + rel_Y};
+    return {short(parentPos[0] + rel_X), short(parentPos[1] + rel_Y)};
 }
